pyproject.toml
[tool.poetry]
name = "study-platform"
version = "0.1.0"
description = ""
authors = ["Lysenko Nikita <nike7cr77@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
black = "^24.10.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
